+++ framework/StandAloneUtilityClasses/TecLibs/3D/Point3D.h
@@ -154,8 +154,8 @@ namespace Kigs
 			// +---------
 			// | Acces Operators
 			// +---------
-			inline const Float& operator[](Int i) const;
-			inline Float& operator[](Int i);
+			inline const Float& operator[](size_t i) const;
+			inline Float& operator[](size_t i);
 
 			Point2D XZ() const { return Point2D(x, z); }
