+++ framework/StandAloneUtilityClasses/TecLibs/3D/Point3DI.h
@@ -71,8 +71,8 @@ namespace Kigs
 			// +---------
 			// | Acces Operators
 			// +---------
-			inline const coordType& operator[](Int i) const;
-			inline coordType& operator[](Int i);
+			inline const coordType& operator[](size_t i) const;
+			inline coordType& operator[](size_t i);
 
 			bool operator==(const Point3DIBase& Other) const
 			{
