+++ framework/StandAloneUtilityClasses/TecLibs/3D/Point3DI.inl.h
@@ -336,14 +336,14 @@ namespace Kigs
         // | Acces Operators
         // +---------
         template<typename coordType>
-        const coordType& Point3DIBase<coordType>::operator[](Int i) const
+        const coordType& Point3DIBase<coordType>::operator[](size_t i) const
         {
             assert(i >= 0 && i < 3);
             return *((&x) + i);
         }
 
         template<typename coordType>
-        coordType& Point3DIBase<coordType>::operator[](Int i)
+        coordType& Point3DIBase<coordType>::operator[](size_t i)
         {
             assert(i >= 0 && i < 3);
             return *((&x) + i);
