+++ framework/StandAloneUtilityClasses/TecLibs/3D/Vector4D.h
@@ -49,8 +49,8 @@ namespace Kigs
 			inline explicit operator const Point3D(void);
 			inline explicit operator const Vector3D(void);
 
-			inline const Float& operator[](Int i) const;
-			inline Float& operator[](Int i);
+			inline const Float& operator[](size_t i) const;
+			inline Float& operator[](size_t i);
 
 			// +---------
 			// | Addition/Substraction
